Extract drop event setup to useTimelineDropHandlers hook
- Created useTimelineDropHandlers hook with drag enter/leave state management
- Extracted handleTrackDragEnter and handleTrackDragLeave functions (32 lines)
- Added drop state management (isDropping, wouldOverlap, dropPosition)
- Reduced timeline-track.tsx from 1,090 to 1,067 lines (9.2% total reduction)
- No TypeScript errors, drop event handling is now reusable
