+++ qcut/apps/web/src/hooks/use-timeline-drop-handlers.ts
@@ -1,4 +1,4 @@
-import { useCallback } from "react";
+import { useCallback, useRef, useState } from "react";
 import { TimelineTrack, TimelineElement } from "@/types/timeline";
 import { SnapPoint } from "@/hooks/use-timeline-snapping";
 
@@ -6,6 +6,11 @@ export interface DropHandlersOptions {
   track: TimelineTrack;
   zoomLevel: number;
   onSnapPointChange?: (snapPoint: SnapPoint | null) => void;
+  getDropSnappedTime: (
+    dropTime: number,
+    elementDuration: number,
+    excludeElementId?: string
+  ) => number;
 }
 
 export interface DropHandlersReturn {
@@ -13,27 +18,60 @@ export interface DropHandlersReturn {
   handleTrackDragLeave: (e: React.DragEvent) => void;
   handleTrackDragOver: (e: React.DragEvent) => void;
   handleTrackDrop: (e: React.DragEvent) => void;
-  isDropTarget: boolean;
+  isDropping: boolean;
   wouldOverlap: boolean;
+  dropPosition: number | null;
 }
 
 /**
  * Custom hook for handling timeline track drop interactions  
  */
 export function useTimelineDropHandlers(options: DropHandlersOptions): DropHandlersReturn {
-  const { track, zoomLevel, onSnapPointChange } = options;
+  const { track, zoomLevel, onSnapPointChange, getDropSnappedTime } = options;
 
-  // Placeholder handlers - will be populated during extraction
+  // Drop state management (extracted from timeline-track.tsx)
+  const [isDropping, setIsDropping] = useState(false);
+  const [dropPosition, setDropPosition] = useState<number | null>(null);
+  const [wouldOverlap, setWouldOverlap] = useState(false);
+  const dragCounterRef = useRef(0);
+
+  // Extracted from timeline-track.tsx (lines 563-577)
   const handleTrackDragEnter = useCallback((e: React.DragEvent) => {
-    // TODO: Extract from timeline-track.tsx (lines ~648-662)
     e.preventDefault();
-    e.stopPropagation();
+
+    const hasTimelineElement = e.dataTransfer.types.includes(
+      "application/x-timeline-element"
+    );
+    const hasMediaItem = e.dataTransfer.types.includes(
+      "application/x-media-item"
+    );
+
+    if (!hasTimelineElement && !hasMediaItem) return;
+
+    dragCounterRef.current++;
+    setIsDropping(true);
   }, []);
 
+  // Extracted from timeline-track.tsx (lines 579-598)
   const handleTrackDragLeave = useCallback((e: React.DragEvent) => {
-    // TODO: Extract from timeline-track.tsx (lines ~664-683)
     e.preventDefault();
-    e.stopPropagation();
+
+    const hasTimelineElement = e.dataTransfer.types.includes(
+      "application/x-timeline-element"
+    );
+    const hasMediaItem = e.dataTransfer.types.includes(
+      "application/x-media-item"
+    );
+
+    if (!hasTimelineElement && !hasMediaItem) return;
+
+    dragCounterRef.current--;
+
+    if (dragCounterRef.current === 0) {
+      setIsDropping(false);
+      setWouldOverlap(false);
+      setDropPosition(null);
+    }
   }, []);
 
   const handleTrackDragOver = useCallback((e: React.DragEvent) => {
@@ -53,7 +91,8 @@ export function useTimelineDropHandlers(options: DropHandlersOptions): DropHandl
     handleTrackDragLeave,
     handleTrackDragOver,
     handleTrackDrop,
-    isDropTarget: false, // TODO: Extract state
-    wouldOverlap: false, // TODO: Extract state
+    isDropping,
+    wouldOverlap,
+    dropPosition,
   };
 }
